Fixed the map and getting location forever
The map wasn’t being created sometimes, turns out google play services
needed to be updated, added a hint to check for updates. Also added
checks everywhere to see if map was null. If the gps was disabled, the
first page stalled on getting location forever. Fixed and also added
logic to get the location again in onresume.