Add check for null coreSettings before using AccountSettings object
Should help with a variety of crashes. Still not sure what's causing the null settings object but in many cases, the app can continue to function without settings. Likely due to a race condition where the app has not fully loaded yet and it's already getting a transaction or a request for an exchange rate.

In some cases when settings fail to load and the code is looking for an account currencyNum settings, just return the CoreAPI default currencynum.